+++ primitives/src/lib.rs
@@ -2,7 +2,10 @@
 #[warn(unused_imports)]
 #[warn(unused_variables)]
 extern crate alloc;
+extern crate core;
 
 pub mod derived_types;
+pub mod error;
+pub mod helpers;
 pub mod types;
 pub mod util;
