+++ primitives/src/types.rs
@@ -1,35 +1,9 @@
-use crate::derived_types;
 use alloc::vec::Vec;
-use core::fmt::{Display, Formatter};
 use ethereum_consensus::{
 	bellatrix::{BeaconBlockHeader, SyncAggregate, SyncCommittee},
-	crypto::PublicKey,
 	domains::DomainType,
-	primitives::{BlsSignature, Hash32, Slot},
-	ssz::ByteVector,
+	primitives::{Hash32, Slot},
 };
-use ssz_rs::{Bitvector, Deserialize, Node, Vector};
-
-#[derive(Debug)]
-pub enum Error {
-	InvalidRoot,
-	InvalidPublicKey,
-	InvalidProof,
-	InvalidBitVec,
-	ErrorConvertingAncestorBlock,
-}
-
-impl Display for Error {
-	fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
-		match self {
-			Error::InvalidRoot => write!(f, "Invalid root",),
-			Error::InvalidPublicKey => write!(f, "Invalid public key",),
-			Error::InvalidProof => write!(f, "Invalid proof",),
-			Error::InvalidBitVec => write!(f, "Invalid bit vec",),
-			Error::ErrorConvertingAncestorBlock => write!(f, "Error deriving ancestor block",),
-		}
-	}
-}
 
 pub const DOMAIN_SYNC_COMMITTEE: DomainType = DomainType::SyncCommittee;
 pub const FINALIZED_ROOT_INDEX: u64 = 52;
@@ -64,34 +38,6 @@ pub struct ExecutionPayloadProof {
 	pub timestamp: u64,
 }
 
-impl TryFrom<derived_types::ExecutionPayloadProof> for ExecutionPayloadProof {
-	type Error = Error;
-	fn try_from(
-		derived_execution_payload_proof: derived_types::ExecutionPayloadProof,
-	) -> Result<Self, Self::Error> {
-		let multi_proof = derived_execution_payload_proof
-			.multi_proof
-			.iter()
-			.map(|proof| Hash32::try_from(proof.as_ref()).map_err(|_| Error::InvalidProof).unwrap())
-			.collect();
-
-		let execution_payload_branch = derived_execution_payload_proof
-			.execution_payload_branch
-			.iter()
-			.map(|proof| Hash32::try_from(proof.as_ref()).map_err(|_| Error::InvalidProof).unwrap())
-			.collect();
-
-		Ok(ExecutionPayloadProof {
-			state_root: Hash32::try_from(derived_execution_payload_proof.state_root.as_slice())
-				.map_err(|_| Error::InvalidRoot)?,
-			block_number: derived_execution_payload_proof.block_number,
-			multi_proof,
-			execution_payload_branch,
-			timestamp: derived_execution_payload_proof.timestamp,
-		})
-	}
-}
-
 /// Holds the neccessary proofs required to verify a header in the `block_roots` field
 /// either in [`BeaconState`] or [`HistoricalBatch`].
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -100,24 +46,6 @@ pub struct BlockRootsProof {
 	pub block_header_index: u64,
 	/// The proof for the header, needed to reconstruct `hash_tree_root(state.block_roots)`
 	pub block_header_branch: Vec<Hash32>,
-}
-
-impl TryFrom<derived_types::BlockRootsProof> for BlockRootsProof {
-	type Error = Error;
-	fn try_from(
-		derived_beacon_block_header: derived_types::BlockRootsProof,
-	) -> Result<Self, Self::Error> {
-		let branch = derived_beacon_block_header
-			.block_header_branch
-			.iter()
-			.map(|proof| Hash32::try_from(proof.as_ref()).map_err(|_| Error::InvalidProof).unwrap())
-			.collect();
-
-		Ok(BlockRootsProof {
-			block_header_index: derived_beacon_block_header.block_header_index,
-			block_header_branch: branch,
-		})
-	}
 }
 
 /// The block header ancestry proof, this is an enum because the header may either exist in
@@ -149,54 +77,6 @@ pub enum AncestryProof {
 	},
 }
 
-impl TryFrom<derived_types::AncestryProof> for AncestryProof {
-	type Error = Error;
-	fn try_from(ancestry_proof: derived_types::AncestryProof) -> Result<Self, Self::Error> {
-		Ok(match ancestry_proof {
-			derived_types::AncestryProof::BlockRoots { block_roots_proof, block_roots_branch } =>
-				AncestryProof::BlockRoots {
-					block_roots_proof: block_roots_proof.try_into()?,
-					block_roots_branch: block_roots_branch
-						.iter()
-						.map(|proof| {
-							Hash32::try_from(proof.as_ref())
-								.map_err(|_| Error::InvalidProof)
-								.unwrap()
-						})
-						.collect(),
-				},
-			derived_types::AncestryProof::HistoricalRoots {
-				block_roots_proof,
-				historical_batch_proof,
-				historical_roots_proof,
-				historical_roots_index,
-				historical_roots_branch,
-			} => AncestryProof::HistoricalRoots {
-				block_roots_proof: block_roots_proof.try_into()?,
-				historical_batch_proof: historical_batch_proof
-					.iter()
-					.map(|proof| {
-						Hash32::try_from(proof.as_ref()).map_err(|_| Error::InvalidProof).unwrap()
-					})
-					.collect(),
-				historical_roots_proof: historical_roots_proof
-					.iter()
-					.map(|proof| {
-						Hash32::try_from(proof.as_ref()).map_err(|_| Error::InvalidProof).unwrap()
-					})
-					.collect(),
-				historical_roots_index,
-				historical_roots_branch: historical_roots_branch
-					.iter()
-					.map(|proof| {
-						Hash32::try_from(proof.as_ref()).map_err(|_| Error::InvalidProof).unwrap()
-					})
-					.collect(),
-			},
-		})
-	}
-}
-
 /// This defines the neccesary data needed to prove ancestor blocks, relative to the finalized
 /// header.
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -209,18 +89,6 @@ pub struct AncestorBlock {
 	pub ancestry_proof: AncestryProof,
 }
 
-impl TryFrom<derived_types::AncestorBlock> for AncestorBlock {
-	type Error = Error;
-	fn try_from(derived_ancestor_block: derived_types::AncestorBlock) -> Result<Self, Self::Error> {
-		let beacon_block_header = construct_beacon_header(derived_ancestor_block.header)?;
-		Ok(AncestorBlock {
-			header: beacon_block_header,
-			execution_payload: derived_ancestor_block.execution_payload.try_into()?,
-			ancestry_proof: derived_ancestor_block.ancestry_proof.try_into()?,
-		})
-	}
-}
-
 /// Holds the latest sync committee as well as an ssz proof for it's existence
 /// in a finalized header.
 #[derive(Debug, Clone, PartialEq, Eq, Default)]
@@ -229,29 +97,6 @@ pub struct SyncCommitteeUpdate<const SYNC_COMMITTEE_SIZE: usize> {
 	pub next_sync_committee: SyncCommittee<SYNC_COMMITTEE_SIZE>,
 	// sync committee, ssz merkle proof.
 	pub next_sync_committee_branch: Vec<Hash32>,
-}
-
-impl<const SYNC_COMMITTEE_SIZE: usize> TryFrom<derived_types::SyncCommitteeUpdate>
-	for SyncCommitteeUpdate<SYNC_COMMITTEE_SIZE>
-{
-	type Error = Error;
-
-	fn try_from(
-		sync_committee_update: derived_types::SyncCommitteeUpdate,
-	) -> Result<Self, Self::Error> {
-		let next_sync_committee =
-			construct_sync_committee(sync_committee_update.next_sync_committee)?;
-		Ok(SyncCommitteeUpdate {
-			next_sync_committee,
-			next_sync_committee_branch: sync_committee_update
-				.next_sync_committee_branch
-				.iter()
-				.map(|proof| {
-					Hash32::try_from(proof.as_ref()).map_err(|_| Error::InvalidProof).unwrap()
-				})
-				.collect(),
-		})
-	}
 }
 
 /// Minimum state required by the light client to validate new sync committee attestations
@@ -266,15 +111,6 @@ pub struct LightClientState<const SYNC_COMMITTEE_SIZE: usize> {
 	pub next_sync_committee: SyncCommittee<SYNC_COMMITTEE_SIZE>,
 }
 
-impl<const SYNC_COMMITTEE_SIZE: usize> TryFrom<derived_types::LightClientState>
-	for LightClientState<SYNC_COMMITTEE_SIZE>
-{
-	type Error = Error;
-	fn try_from(state: derived_types::LightClientState) -> Result<Self, Self::Error> {
-		construct_light_client_state(state)
-	}
-}
-
 /// Finalized header proof
 #[derive(Debug, Clone, PartialEq, Eq, Default)]
 pub struct FinalityProof {
@@ -282,22 +118,6 @@ pub struct FinalityProof {
 	pub epoch: u64,
 	/// Finalized header proof
 	pub finality_branch: Vec<Hash32>,
-}
-
-impl TryFrom<derived_types::FinalityProof> for FinalityProof {
-	type Error = Error;
-	fn try_from(derived_finality_proof: derived_types::FinalityProof) -> Result<Self, Self::Error> {
-		Ok(FinalityProof {
-			epoch: derived_finality_proof.epoch,
-			finality_branch: derived_finality_proof
-				.finality_branch
-				.iter()
-				.map(|proof| {
-					Hash32::try_from(proof.as_ref()).map_err(|_| Error::InvalidProof).unwrap()
-				})
-				.collect(),
-		})
-	}
 }
 
 /// Data required to advance the state of the light client.
@@ -319,122 +139,4 @@ pub struct LightClientUpdate<const SYNC_COMMITTEE_SIZE: usize> {
 	pub signature_slot: Slot,
 	/// ancestors of the finalized block to be verified, may be empty.
 	pub ancestor_blocks: Vec<AncestorBlock>,
-}
-
-impl<const SYNC_COMMITTEE_SIZE: usize> TryFrom<derived_types::LightClientUpdate>
-	for LightClientUpdate<SYNC_COMMITTEE_SIZE>
-{
-	type Error = Error;
-	fn try_from(derived_update: derived_types::LightClientUpdate) -> Result<Self, Self::Error> {
-		construct_light_client_update(derived_update)
-	}
-}
-
-fn construct_beacon_header(
-	derived_header: derived_types::BeaconBlockHeader,
-) -> Result<BeaconBlockHeader, Error> {
-	let finalized_header = BeaconBlockHeader {
-		slot: derived_header.slot,
-		proposer_index: derived_header.proposer_index as usize,
-		parent_root: Node::from_bytes(
-			derived_header.parent_root.as_ref().try_into().map_err(|_| Error::InvalidRoot)?,
-		),
-		state_root: Node::from_bytes(
-			derived_header.state_root.as_ref().try_into().map_err(|_| Error::InvalidRoot)?,
-		),
-		body_root: Node::from_bytes(
-			derived_header.body_root.as_ref().try_into().map_err(|_| Error::InvalidRoot)?,
-		),
-	};
-
-	Ok(finalized_header)
-}
-
-fn construct_sync_committee<const SYNC_COMMITTEE_SIZE: usize>(
-	derived_sync_committee: derived_types::SyncCommittee,
-) -> Result<SyncCommittee<SYNC_COMMITTEE_SIZE>, Error> {
-	let public_keys_vector: Vec<PublicKey> = derived_sync_committee
-		.public_keys
-		.iter()
-		.map(|public_key| {
-			PublicKey::try_from(public_key.as_slice())
-				.map_err(|_| Error::InvalidPublicKey)
-				.unwrap()
-		})
-		.collect();
-	let sync_committee = SyncCommittee {
-		public_keys: Vector::try_from(public_keys_vector).unwrap(),
-		aggregate_public_key: PublicKey::try_from(
-			derived_sync_committee.aggregate_public_key.as_slice(),
-		)
-		.map_err(|_| Error::InvalidPublicKey)?,
-	};
-
-	Ok(sync_committee)
-}
-
-fn construct_sync_aggregate<const SYNC_COMMITTEE_SIZE: usize>(
-	derived_sync_aggregate: derived_types::SyncAggregate,
-) -> Result<SyncAggregate<SYNC_COMMITTEE_SIZE>, Error> {
-	let derived_sync_committee_bits = derived_sync_aggregate.sync_committee_bits;
-	let bit_vector = Bitvector::<SYNC_COMMITTEE_SIZE>::deserialize(&derived_sync_committee_bits)
-		.map_err(|_| Error::InvalidBitVec)?;
-
-	let sync_aggregate = SyncAggregate {
-		sync_committee_bits: bit_vector,
-		sync_committee_signature: BlsSignature::try_from(
-			derived_sync_aggregate.sync_committee_signature.as_ref(),
-		)
-		.map_err(|_| Error::InvalidPublicKey)?,
-	};
-
-	Ok(sync_aggregate)
-}
-
-fn construct_light_client_state<const SYNC_COMMITTEE_SIZE: usize>(
-	state: derived_types::LightClientState,
-) -> Result<LightClientState<SYNC_COMMITTEE_SIZE>, Error> {
-	let finalized_header = construct_beacon_header(state.finalized_header)?;
-
-	let current_sync_committee = construct_sync_committee(state.current_sync_committee.clone())?;
-	let next_sync_committee = construct_sync_committee(state.next_sync_committee)?;
-
-	Ok(LightClientState {
-		finalized_header,
-		latest_finalized_epoch: state.latest_finalized_epoch,
-		current_sync_committee,
-		next_sync_committee,
-	})
-}
-
-fn construct_light_client_update<const SYNC_COMMITTEE_SIZE: usize>(
-	derived_update: derived_types::LightClientUpdate,
-) -> Result<LightClientUpdate<SYNC_COMMITTEE_SIZE>, Error> {
-	let sync_committee_update_option: Option<SyncCommitteeUpdate<SYNC_COMMITTEE_SIZE>>;
-
-	match derived_update.sync_committee_update {
-		Some(sync_committee_update) =>
-			sync_committee_update_option = Some(sync_committee_update.try_into()?),
-		None => sync_committee_update_option = None,
-	}
-	Ok(LightClientUpdate {
-		attested_header: construct_beacon_header(derived_update.attested_header)?,
-		sync_committee_update: sync_committee_update_option,
-		finalized_header: construct_beacon_header(derived_update.finalized_header)?,
-		execution_payload: derived_update.execution_payload.try_into()?,
-		finality_proof: derived_update.finality_proof.try_into()?,
-		sync_aggregate: construct_sync_aggregate(derived_update.sync_aggregate)?,
-		signature_slot: derived_update.signature_slot,
-		ancestor_blocks: derived_update
-			.ancestor_blocks
-			.iter()
-			.map(|ancestor_block| {
-				ancestor_block
-					.clone()
-					.try_into()
-					.map_err(|_| Error::ErrorConvertingAncestorBlock)
-					.unwrap()
-			})
-			.collect(),
-	})
 }
